Bug 72453 - Modify size_t and uint32_t for gfx layers for arm64. r=Eastern

+++ gfx/layers/ipc/GonkNativeHandleUtils.cpp
@@ -51,7 +51,7 @@ bool
 ParamTraits<GonkNativeHandle>::Read(const Message* aMsg,
                                PickleIterator* aIter, paramType* aResult)
 {
-  size_t nbytes;
+  uint32_t nbytes;
   // TODO: verify this is correct after bug 1525199.
   if (!aMsg->ReadUInt32(aIter, &nbytes)) {
     return false;
